Убрал избыточное димирование и размытие

+++ include/effects.h
@@ -1029,7 +1029,7 @@ private:
   uint8_t waveCount = 1;
   const uint8_t waveScale = 256 / WIDTH;
   uint8_t waveRotation;
-  uint8_t whue;
+  float whue;
   float waveTheta;
   bool wavesRoutine(CRGB *leds, EffectWorker *param);
   void palettemap(std::vector<PGMPalette*> &_pals, const uint8_t _val, const uint8_t _min, const uint8_t _max) override;
